LPS-56566 Move to specialized class
