+++ modules/apps/dynamic-data-lists/dynamic-data-lists-service/src/com/liferay/dynamic/data/lists/util/impl/DDLImpl.java
@@ -14,8 +14,6 @@
 
 package com.liferay.dynamic.data.lists.util.impl;
 
-import com.liferay.dynamic.data.lists.configuration.DDLServiceConfigurationUtil;
-import com.liferay.dynamic.data.lists.configuration.DDLServiceConfigurationValues;
 import com.liferay.dynamic.data.lists.exception.NoSuchRecordException;
 import com.liferay.dynamic.data.lists.model.DDLRecord;
 import com.liferay.dynamic.data.lists.model.DDLRecordConstants;
@@ -25,9 +23,6 @@ import com.liferay.dynamic.data.lists.service.DDLRecordLocalServiceUtil;
 import com.liferay.dynamic.data.lists.service.DDLRecordServiceUtil;
 import com.liferay.dynamic.data.lists.service.DDLRecordSetLocalServiceUtil;
 import com.liferay.dynamic.data.lists.util.DDL;
-import com.liferay.dynamic.data.lists.util.DDLConstants;
-import com.liferay.portal.kernel.configuration.Filter;
-import com.liferay.portal.kernel.io.unsync.UnsyncStringWriter;
 import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
@@ -38,12 +33,6 @@ import com.liferay.portal.kernel.repository.model.FileEntry;
 import com.liferay.portal.kernel.search.Hits;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistryUtil;
-import com.liferay.portal.kernel.template.TemplateConstants;
-import com.liferay.portal.kernel.template.TemplateHandler;
-import com.liferay.portal.kernel.template.TemplateHandlerRegistryUtil;
-import com.liferay.portal.kernel.template.TemplateManager;
-import com.liferay.portal.kernel.template.TemplateManagerUtil;
-import com.liferay.portal.kernel.util.Constants;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.LocaleThreadLocal;
 import com.liferay.portal.kernel.util.LocaleUtil;
@@ -52,14 +41,10 @@ import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.service.LayoutServiceUtil;
 import com.liferay.portal.service.ServiceContext;
-import com.liferay.portal.templateparser.Transformer;
-import com.liferay.portal.theme.ThemeDisplay;
 import com.liferay.portlet.documentlibrary.service.DLAppLocalServiceUtil;
 import com.liferay.portlet.dynamicdatamapping.model.DDMFormField;
 import com.liferay.portlet.dynamicdatamapping.model.DDMStructure;
-import com.liferay.portlet.dynamicdatamapping.model.DDMTemplate;
 import com.liferay.portlet.dynamicdatamapping.model.LocalizedValue;
-import com.liferay.portlet.dynamicdatamapping.service.DDMTemplateLocalServiceUtil;
 import com.liferay.portlet.dynamicdatamapping.storage.DDMFormValues;
 import com.liferay.portlet.dynamicdatamapping.storage.Field;
 import com.liferay.portlet.dynamicdatamapping.storage.Fields;
@@ -70,15 +55,10 @@ import com.liferay.portlet.dynamicdatamapping.util.DDMUtil;
 
 import java.util.ArrayList;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Locale;
-import java.util.Map;
-import java.util.Set;
 
 import javax.portlet.PortletPreferences;
-import javax.portlet.RenderRequest;
-import javax.portlet.RenderResponse;
 
 import javax.servlet.http.HttpServletRequest;
 
@@ -314,62 +294,6 @@ public class DDLImpl implements DDL {
 		return jsonArray;
 	}
 
-	@Override
-	public String getTemplateContent(
-			long ddmTemplateId, DDLRecordSet recordSet,
-			ThemeDisplay themeDisplay, RenderRequest renderRequest,
-			RenderResponse renderResponse)
-		throws Exception {
-
-		Transformer transformer = TransformerHolder.getTransformer();
-
-		Map<String, Object> contextObjects = new HashMap<>();
-
-		contextObjects.put(
-			DDLConstants.RESERVED_DDM_STRUCTURE_ID,
-			recordSet.getDDMStructureId());
-		contextObjects.put(
-			DDLConstants.RESERVED_DDM_TEMPLATE_ID, ddmTemplateId);
-		contextObjects.put(
-			DDLConstants.RESERVED_RECORD_SET_DESCRIPTION,
-			recordSet.getDescription(themeDisplay.getLocale()));
-		contextObjects.put(
-			DDLConstants.RESERVED_RECORD_SET_ID, recordSet.getRecordSetId());
-		contextObjects.put(
-			DDLConstants.RESERVED_RECORD_SET_NAME,
-			recordSet.getName(themeDisplay.getLocale()));
-		contextObjects.put(TemplateConstants.TEMPLATE_ID, ddmTemplateId);
-
-		String viewMode = Constants.VIEW;
-
-		if (renderRequest != null) {
-			viewMode = ParamUtil.getString(
-				renderRequest, "viewMode", Constants.VIEW);
-		}
-
-		contextObjects.put("viewMode", viewMode);
-
-		DDMTemplate ddmTemplate = DDMTemplateLocalServiceUtil.getTemplate(
-			ddmTemplateId);
-
-		contextObjects.put(
-			TemplateConstants.CLASS_NAME_ID, ddmTemplate.getClassNameId());
-
-		TemplateManager templateManager =
-			TemplateManagerUtil.getTemplateManager(ddmTemplate.getLanguage());
-
-		TemplateHandler templateHandler =
-			TemplateHandlerRegistryUtil.getTemplateHandler(
-				DDLRecordSet.class.getName());
-
-		templateManager.addContextObjects(
-			contextObjects, templateHandler.getCustomContextObjects());
-
-		return transformer.transform(
-			themeDisplay, contextObjects, ddmTemplate.getScript(),
-			ddmTemplate.getLanguage(), new UnsyncStringWriter());
-	}
-
 	/**
 	 * @deprecated As of 7.0.0, with no direct replacement
 	 */
@@ -486,36 +410,5 @@ public class DDLImpl implements DDL {
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(DDLImpl.class);
-
-	private static class TransformerHolder {
-
-		public static Transformer getTransformer() {
-			return _transformer;
-		}
-
-		private static final Transformer _transformer = new Transformer(
-			DDLServiceConfigurationValues.DYNAMIC_DATA_LISTS_ERROR_TEMPLATE,
-			true) {
-
-			@Override
-			protected void setErrorTemplateIds(
-				String errorTemplatePropertyKey) {
-
-				Set<String> langTypes =
-					TemplateManagerUtil.getTemplateManagerNames();
-
-				for (String langType : langTypes) {
-					String errorTemplateId = DDLServiceConfigurationUtil.get(
-						errorTemplatePropertyKey, new Filter(langType));
-
-					if (Validator.isNotNull(errorTemplateId)) {
-						_errorTemplateIds.put(langType, errorTemplateId);
-					}
-				}
-			}
-
-		};
-
-	}
 
 }
